Respect the -sb and -ss arguments.
Current modification consists at removing the code that forces remu.
A probable better way, would be to allow the remu force if and only if -sb or -ss are not provided.
This behavior would be better than the current one (always forced) or the one I suggest in this patch (never forced).